bb: Rename BBTagInfoMap2 parts to BBLV_Metadata
