+++ bb/src/BBLV_Metadata.h
