+++ bb/src/xfer.cc
@@ -37,6 +37,7 @@ namespace bs  = boost::system;
 #include "bbinternal.h"
 #include "bbio_regular.h"
 #include "bbio_BSCFS.h"
+#include "BBLV_Metadata.h"
 #include "bbserver_flightlog.h"
 #include "BBTagID.h"
 #include "BBTransferDef.h"
@@ -45,7 +46,6 @@ namespace bs  = boost::system;
 #include "BBTagInfo.h"
 #include "BBTagInfo2.h"
 #include "BBTagInfoMap.h"
-#include "BBTagInfoMap2.h"
 #include "ContribIdFile.h"
 #include "ExtentInfo.h"
 #include "identity.h"
